fix(exams): 受験回数が文字列連結になる不具合と展開ボタンが効かない不具合を修正
- PostgreSQLの count()/sum()/max() は bigint/numeric のため node-postgres が
  文字列で返し、受験回数の合計が "01000" のような連結になっていた。
  API側で ::int にキャストし、UI側でも Number() で二重に保護した
- ジャンルの展開ボタンが、ジャンムが1つのとき常に開く条件になっていて
  クリックしても反応しなかった。既定を開いた状態にし、
  クリックで閉じられる方式に変更(検索中は常に開く)

+++ server/src/routes/exams.ts
@@ -35,12 +35,12 @@ examsRouter.get('/exams', async (req, res) => {
   const { rows } = await pool.query(
     `SELECT e.id, e.name, e.description, e.group_name, e.pass_score, e.time_limit_min,
             e.sort_order, e.is_published,
-            (SELECT count(*) FROM exam_questions q WHERE q.exam_id = e.id)          AS question_count,
-            (SELECT COALESCE(sum(q.points), 0) FROM exam_questions q WHERE q.exam_id = e.id) AS total_points,
+            (SELECT count(*) FROM exam_questions q WHERE q.exam_id = e.id)::int      AS question_count,
+            (SELECT COALESCE(sum(q.points), 0) FROM exam_questions q WHERE q.exam_id = e.id)::int AS total_points,
             (SELECT count(*) FROM exam_attempts a
-              WHERE a.exam_id = e.id AND a.user_id = $1 AND a.status = 'submitted') AS my_attempts,
+              WHERE a.exam_id = e.id AND a.user_id = $1 AND a.status = 'submitted')::int AS my_attempts,
             (SELECT max(a.earned_points) FROM exam_attempts a
-              WHERE a.exam_id = e.id AND a.user_id = $1 AND a.status = 'submitted') AS my_best_score
+              WHERE a.exam_id = e.id AND a.user_id = $1 AND a.status = 'submitted')::int AS my_best_score
        FROM exams e
       WHERE e.is_published OR $2 = 'board'
       ORDER BY e.group_name, e.sort_order, e.id`,
